refactor: FoodList class 위치 변경
